conversion of types numbers

+++ level10/lesson04/task05/Solution.java
@@ -12,8 +12,8 @@ public class Solution
     public static void main(String[] args)
     {
         int a = (byte)44;
-        int b = (byte)300;
-        short c = (byte)(b - a);
+        int b = 300;
+        short c = (short) (b - a);
         System.out.println(c);
     }
 }
